+++ CHANGELOG.md
@@ -2,17 +2,20 @@
 
 Go-Common is the library of common functions for Tidepool's Go-based applications
 
+## Unreleased
+- PT-1383 Add portal-api client to fetch the PatientConfig
+
 ## 0.4.0 - 2020-04-14
 ### Added
 - Complete Getekeeper client with missing route
 
 ## 0.3.0 - 2019-10-17
 ### Changed
-- [PT-727] Add the versioning info in the status object
+- PT-727 Add the versioning info in the status object
 
 ## 0.2.0 - 2019-10-09
 ### Added
-- [PT-582] Merge from upstream Tidepool v0.4.1
+- PT-582 Merge from upstream Tidepool v0.4.1
 
 ## 0.1.0 - 2019-05-29
 ### Added